+++ src/main/java/com/ltk/forum/repository/StatusRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ltk.forum.model.Status;
 
 
-public interface StatusRepo extends JpaRepository<Status, Integer> {
+public interface StatusRepo extends JpaRepository<Status, Long> {
 
 	Status findOneById(Long id);
 	Status findOneByStatusCode(String statusCode);
