+++ src/main/java/com/ltk/forum/repository/TypeOfPostRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ltk.forum.model.TypeOfPost;
 
 
-public interface TypeOfPostRepo extends JpaRepository<TypeOfPost, Integer>{
+public interface TypeOfPostRepo extends JpaRepository<TypeOfPost, Long>{
 
 	TypeOfPost findOneById(Long id);
 	TypeOfPost findOneByTypeCode(String typeCode);
