+++ src/main/java/com/ltk/forum/repository/TypeOfReportRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ltk.forum.model.TypeOfReport;
 
 
-public interface TypeOfReportRepo extends JpaRepository<TypeOfReport, Integer> {
+public interface TypeOfReportRepo extends JpaRepository<TypeOfReport, Long> {
 
 	TypeOfReport findOneById(Long id);
 	TypeOfReport findOneByTypeCode(String typeCode);
